Update ASPC_Engine.ps1 - xternal mail fix 2
The first fix added a check for null. IT appears that it was a blank space causing the issue. Removed the null / blank check and added (if !$variable) this will check if the string is empty or null, if so it writes to the log file. If not, it will send an email.